webMAN MOD 1.45.01
- New USER/ADMIN security mode:
* Several functions are limited while the plugin is in USER mode
* FTP server and some remote functions are temporarily disabled
* Added combo L2+R2+TRIANGLE to toggle USER/ADMIN mode
* Added /admin.ps3?enable&pwd=<pwd> to change mode
* Added pwd=<pwd> parameter to setup, install, restart, shutdown

- Fixed fan controller paused during PKG installation via combo

- Fixed Remove SingStar Icon (thanks to bguerville)
- Fixed option Remove SingStar Icon couldn't be changed

- Fixed PSP games not mounted due fail copying icon
- Fixed dup. games listing on full edition if net srv is unavailable
- Added option to mute sound of mounted game when listed on XMB

- Mounting from XMB always close folders (even if autoplay is off)
(This option can be disabled unchecking PLAY DISC combo)
- Led now always blink during game mounting

- Auto Power Off is now an optional module (added flag)
- Added option to prevent auto-power off on ftp, install, dl
(in previous version the option was enabled through USB polling)
- Removed Auto Power Off from Lite edition to keep size <192KB

- Some system combos now work only on XMB (e.g togglers)
- Several loops now exit immediately if working = 0
- Set explicit buffer size in ftp server
- Set ps1_netemu now removes mapping instead remap to origin

- Fixed /net games not listed in Last Game list of File Manager
- Added 3 second timeout to ps3netsrv connections
- Fixed crash still happening on ps3netsrv (Windows version)

+++ include/combos.h
@@ -11,10 +11,16 @@
  NEXT GAME    : SELECT+R1                       *or* Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_select_r1
  UMNT_GAME    : SELECT+O (unmount)              *or* Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_select_circle
 
+ RESERVED     : SELECT+TRIANGLE                 *or* Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_select_triangle
+
+ EXT GAME DATA: SELECT+□                        *or* Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_select_square
+ MOUNT net0/  : SELECT+R2+□
+ MOUNT net1/  : SELECT+L2+□
+
  SHUTDOWN     : L3+R2+X
- SHUTDOWN  *2 : L3+R1+X (vsh shutdown)
+ SHUTDOWN  *2 : L3+R1+X (vsh shutdown) <- alternative shutdown method
  RESTART      : L3+R2+O (lpar restart)
- RESTART   *2 : L3+R1+O (vsh restart)
+ RESTART   *2 : L3+R1+O (vsh restart)  <- alternative restart method
 
  FAN CNTRL    : L3+R2+START  (enable/disable fancontrol)
  SHOW TEMP    : SELECT+START (SELECT+START+R2 will show only copy progress) / SELECT+R3 (if rec video flag is disabled)
@@ -26,13 +32,11 @@
  REC VIDEO VAL: SELECT+R3+R2  Change value of video rec setting
  XMB SCRNSHOT : L2+R2+SELECT+START
 
+ USER/ADMIN   : L2+R2+TRIANGLE                  *or* Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_l2_r2_triangle
+
  SYSCALLS     : R2+TRIANGLE                     *or* Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_r2_triangle
  SHOW IDPS    : R2+O  (Abort copy/fix process)  *or* Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_r2_circle
  OFFLINE MODE : R2+□                            *or* Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_r2_square
-
- EXT GAME DATA: SELECT+□                        *or* Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_select_square
- MOUNT net0/  : SELECT+R2+□
- MOUNT net1/  : SELECT+L2+□
 
  QUICK INSTALL: SELECT+R2+O                     *or* Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_select_r2_circle
 
@@ -44,7 +48,7 @@
  Normal Mode Switcher : L3+L2+O
  DEBUG  Menu Switcher : L3+L2+X
 
- Skip auto-mount   : L2+R2                           Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_l2_r2 (not overriden)
+ SKIP AUTO-MOUNT   : L2+R2  (at startup only)   *or* Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_l2_r2
 
  Open File Manager : L2+R2+O                    *or* Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_l2_r2_circle
  Open Games List   : L2+R2+R1+O                 *or* Custom Combo -> /dev_hdd0/tmp/wm_combo/wm_custom_l2_r2_r1_circle
@@ -83,7 +87,7 @@
 				{
 					if(!(webman_config->combo2 & PLAY_DISC) && (data.button[CELL_PAD_BTN_OFFSET_DIGITAL1] == CELL_PAD_CTRL_START) && (data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] == CELL_PAD_CTRL_L2))
 					{
-						char category[16] = "game", seg_name[80]; sprintf(seg_name, "seg_device");
+						char category[16], seg_name[40]; *category = *seg_name = NULL;
 						launch_disc(category, seg_name); // L2+START
 						break;
 					}
@@ -93,15 +97,17 @@
 						if( !(webman_config->combo2 & (EXTGAMDAT | MOUNTNET0 | MOUNTNET1))         // Toggle External Game Data
 							&& (data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] & CELL_PAD_CTRL_SQUARE)) // SELECT+SQUARE
 						{
-#ifndef LITE_EDITION
+#ifdef COBRA_ONLY
+ #ifndef LITE_EDITION
 							if(!(webman_config->combo2 & MOUNTNET0) &&
 								(data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] & CELL_PAD_CTRL_R2))
-							{if(webman_config->netp[0] && webman_config->neth[0][0]) mount_with_mm((char*)"/net0", 1);} // SELECT+SQUARE+R2
+							{if(is_netsrv_enabled(0)) mount_with_mm((char*)"/net0", 1);} // SELECT+SQUARE+R2 / SELECT+R2+SQUARE
 							else
 							if(!(webman_config->combo2 & MOUNTNET1) &&
 								(data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] & CELL_PAD_CTRL_L2))
-							{if(webman_config->netp[1] && webman_config->neth[1][0]) mount_with_mm((char*)"/net1", 1);} // SELECT+SQUARE+L2
+							{if(is_netsrv_enabled(1)) mount_with_mm((char*)"/net1", 1);} // SELECT+SQUARE+L2 / SELECT+L2+SQUARE
 							else
+ #endif
 #endif
 							{
 #ifdef WM_CUSTOM_COMBO
@@ -123,6 +129,13 @@
 							(data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] == (CELL_PAD_CTRL_L2 | CELL_PAD_CTRL_R2)) // SELECT+L3+L2+R2
 							)
 						{
+							//// startup time /////
+							CellRtcTick pTick; cellRtcGetCurrentTick(&pTick);
+							u32 ss = (u32)((pTick.tick - rTick.tick)/1000000);
+							///////////////////////
+
+							if((!sys_admin && (ss > 60)) || IS_INGAME) continue;
+
 							cellFsUnlink("/dev_hdd0/boot_plugins.txt");
 							goto reboot; // vsh reboot
 						}
@@ -131,6 +144,7 @@
 							(data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] == (CELL_PAD_CTRL_L2 | CELL_PAD_CTRL_R2)) // SELECT+R3+L2+R2
 							)
 						{
+							if(!sys_admin || IS_INGAME) continue;
 #ifndef ENGLISH_ONLY
 							char STR_RMVWMCFG[96];//		= "webMAN config reset in progress...";
 							char STR_RMVWMCFGOK[112];//	= "Done! Restart within 3 seconds";
@@ -454,12 +468,9 @@ show_popup:
 							else
 #endif
 							{
-								led(GREEN, BLINK_FAST);
 								mount_with_mm((char*)"_prev", 1);
-								sys_timer_sleep(2);
-								led(GREEN, ON);
 
-								n = 10;
+								n = 0;
 								break;
 							}
 						}
@@ -471,12 +482,9 @@ show_popup:
 							else
 #endif
 							{
- 								led(GREEN, BLINK_FAST);
 								mount_with_mm((char*)"_next", 1);
-								sys_timer_sleep(2);
-								led(GREEN, ON);
 
-								n = 10;
+								n = 0;
 								break;
 							}
 						}
@@ -488,7 +496,11 @@ show_popup:
 							if(do_custom_combo("select_r2_circle")) break;
 							else
 #endif
-							installPKG_combo(msg);
+							if(!install_in_progress && IS_ON_XMB)
+							{
+								sys_ppu_thread_t thread_id;
+								sys_ppu_thread_create(&thread_id, installPKG_combo_thread, NULL, THREAD_PRIO, THREAD_STACK_SIZE_8KB, SYS_PPU_THREAD_CREATE_NORMAL, THREAD_NAME_INSTALLPKG);
+							}
 
 							n = 0;
 							break;
@@ -510,7 +522,7 @@ show_popup:
 						else
 						if(!(webman_config->combo & UMNT_GAME) && (data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] == CELL_PAD_CTRL_TRIANGLE) ) // SELECT+TRIANGLE
 						{
-							if(do_custom_combo("select_triangle")) break;
+							if(do_custom_combo("select_triangle")) break; // RESERVED
 						}
 #endif
 					}
@@ -575,7 +587,7 @@ show_popup:
 
 							sys_ppu_thread_exit(0);
 						}
-						else
+
 						if(!(webman_config->combo & RESTARTPS) && (data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] & CELL_PAD_CTRL_CIRCLE)) // L3+R1+O (vsh restart)
 						{
 							// vsh reboot
@@ -587,6 +599,26 @@ show_popup:
 							sys_ppu_thread_exit(0);
 						}
 					}
+#ifdef SYS_ADMIN_MODE
+					else
+					if((webman_config->combo & SYS_ADMIN) && (data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] == (CELL_PAD_CTRL_L2 | CELL_PAD_CTRL_R2 | CELL_PAD_CTRL_TRIANGLE) )) // L2+R2+TRIANGLE Toggle user/admin mode
+					{
+ #ifdef WM_CUSTOM_COMBO
+						if(do_custom_combo("l2_r2_triangle")) break;
+						else
+ #endif
+						{
+							sys_admin ^= 1, pwd_tries = 0;
+
+							sprintf(msg, "ADMIN %s", sys_admin ? STR_ENABLED : STR_DISABLED);
+							show_msg(msg);
+
+							if(sys_admin) { BEEP1 } else { BEEP2 }
+						}
+						n = 0;
+						break;
+					}
+#endif
 					else
 					if(data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] & CELL_PAD_CTRL_R2)
 					{
@@ -617,7 +649,7 @@ show_popup:
 						else
 						if((copy_in_progress || fix_in_progress) && (data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] & CELL_PAD_CTRL_CIRCLE)) // R2+O Abort copy process
 						{
-							fix_aborted = copy_aborted =true;
+							fix_aborted = copy_aborted = true;
 						}
 						else
 						if(!(webman_config->combo & DISABLESH) && (data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] & CELL_PAD_CTRL_TRIANGLE) ) // R2+TRIANGLE Disable CFW Sycalls
@@ -667,7 +699,7 @@ show_popup:
 						else
 						if(!(webman_config->combo & SHOW_IDPS) && (data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] == (CELL_PAD_CTRL_L2 | CELL_PAD_CTRL_R2))) // L2+R2
 						{
-							if(do_custom_combo("l2_r2") == false) continue;
+							if(do_custom_combo("l2_r2") == false) continue; // RESERVED
 						}
 #endif
 						n = 0;
@@ -676,6 +708,7 @@ show_popup:
 					else
 					if((data.button[CELL_PAD_BTN_OFFSET_DIGITAL1] & CELL_PAD_CTRL_L3) && (data.button[CELL_PAD_BTN_OFFSET_DIGITAL2] & CELL_PAD_CTRL_L2))
 					{
+						if(!sys_admin || IS_INGAME) continue;
 #ifdef COBRA_ONLY
  #ifndef LITE_EDITION
 						if(!(webman_config->combo & DISACOBRA)
@@ -727,7 +760,8 @@ reboot:
 				}
 
 			}
-			//sys_timer_sleep(step);
+
+			if(!working) break;
 			sys_timer_usleep(300000);
 
 			if(show_persistent_popup)
@@ -738,4 +772,4 @@ reboot:
 			}
 		}
 
-		if(n < 10) sys_timer_usleep((12-n) * 150000);
+		if(working && (n < 10)) sys_timer_usleep((12 - n) * 150000);
